Added remote self host feature
Squashed commit of the following:

commit 717c4a0a4f76d85649546cbae02fd6395009753d
Author: Nico Prananta <311343+nicnocquee@users.noreply.github.com>
Date:   Mon Nov 11 23:33:05 2024 +0100

    fix: add example environment file with essential configuration variables for local development setup

commit 4c773601a2408797e9b9788fa3123610424d0481
Author: Nico Prananta <311343+nicnocquee@users.noreply.github.com>
Date:   Mon Nov 11 23:29:25 2024 +0100

    fix: update post structure to use JSONB format and streamline metadata fetching with new environment variables in deployment workflow

commit dc690a25b77f06655279b3e506ffce5ac09dff8a
Author: Nico Prananta <311343+nicnocquee@users.noreply.github.com>
Date:   Mon Nov 11 01:51:27 2024 +0100

    fix: restructure post handling by introducing getPostData to simplify data preparation before post creation in PostScheduler component

commit 98e618a7584b1a6b37698674c2bf02b267a2b33b
Author: Nico Prananta <311343+nicnocquee@users.noreply.github.com>
Date:   Mon Nov 11 01:10:49 2024 +0100

    fix: add cron service to Docker Compose for regular API requests to check posts every minute

commit d61e5e5153016e66ae4a9405ec5f6da5f4cf88f3
Author: Nico Prananta <311343+nicnocquee@users.noreply.github.com>
Date:   Mon Nov 11 00:11:12 2024 +0100

    fix: add fetching posts at regular intervals in ScheduledPosts component to ensure updates every minute

commit e60e0c55dae01ab7ef8b44022376579b1aa49905
Author: Nico Prananta <311343+nicnocquee@users.noreply.github.com>
Date:   Mon Nov 11 00:07:26 2024 +0100

    fix: sort fetched posts by scheduled date in descending order before setting state in ScheduledPosts component

commit 62c2cdb8b5f60e2fe9c03c2f88cfa0fa934cfc22
Author: Nico Prananta <311343+nicnocquee@users.noreply.github.com>
Date:   Mon Nov 11 00:03:44 2024 +0100

    fix: update dev script to use local storage mode and conditionally render OfflineInfo based on storage configuration

commit c72815a418fbb6cfc4f6587eb0c0d629bd040f78
Author: Nico Prananta <311343+nicnocquee@users.noreply.github.com>
Date:   Sun Nov 10 23:48:55 2024 +0100

    chore: refactor docker-compose to use environment variables for configuration; add cron service to check posts regularly

commit 7931e63ccdb654b8f21e34cde282eb8bae696405
Author: Nico Prananta <311343+nicnocquee@users.noreply.github.com>
Date:   Sun Nov 10 23:36:48 2024 +0100

    feat: implement API authentication setup and login forms; enhance express routes for credential management and integration with components

commit 3e0ae1ad03895e9f655cba3c66f284bc06879942
Author: Nico Prananta <311343+nicnocquee@users.noreply.github.com>
Date:   Sat Nov 9 11:49:11 2024 +0100

    enhance API with dotenv for environment variables, add endpoints for credentials & improve local storage integration in components

commit 1123bb0f486451c8c47d1af74417b5189792f53a
Author: Nico Prananta <311343+nicnocquee@users.noreply.github.com>
Date:   Sat Nov 9 09:24:25 2024 +0100

    update title in index.html to reflect rebranding from "Blusky Post Scheduler" to "Bluesky Later - Bluesky Post Scheduler"

commit 7ae3a0fe39a827ab5e7fbfcf96ddff93535b254d
Author: Nico Prananta <311343+nicnocquee@users.noreply.github.com>
Date:   Sat Nov 9 01:43:48 2024 +0100

    update project configuration: add .env to .gitignore, add Docker support for PostgreSQL, and enhance package.json scripts for development

commit cddecca18aba005516b436e0ac24ee361361933f
Author: Nico Prananta <311343+nicnocquee@users.noreply.github.com>
Date:   Sat Nov 9 01:09:17 2024 +0100

    refactor PostScheduler and ScheduledPosts components to use unified db methods for creating and deleting posts, enhancing code consistency

commit 53b8d7e97a0785b40b25515796927e079925f8bb
Author: Nico Prananta <311343+nicnocquee@users.noreply.github.com>
Date:   Sat Nov 9 00:44:39 2024 +0100

    add Docker, Nginx, and API setup with PostgreSQL for Bluesky Scheduler application, including local and remote database handling

+++ src/app.tsx
@@ -1,39 +1,44 @@
-import { useEffect, useState } from "react";
+import { useEffect } from "react";
 import { LoginForm } from "./components/login-form";
 import { PostScheduler } from "./components/post-scheduler";
 import { ScheduledPosts } from "./components/scheduled-posts";
-import { getStoredCredentials } from "./lib/bluesky";
 import { Toaster } from "react-hot-toast";
 import { db } from "./lib/db";
 import { Footer } from "./components/footer";
+import { SetupForm } from "./components/setup-form";
+import { ApiLoginForm } from "./components/api-login-form";
+import { useAuth } from "./components/use-auth";
 
 function App() {
-  const [identifier, setIdentifier] = useState<string | undefined>();
-  const [isLoading, setIsLoading] = useState(true);
+  const {
+    identifier,
+    isLoading,
+    hasApiCredentials,
+    isApiAuthenticated,
+    updateIdentifier,
+    updateApiAuth,
+  } = useAuth();
 
   useEffect(() => {
-    const checkAuth = async () => {
-      const creds = await getStoredCredentials();
-      setIdentifier(creds?.identifier);
-      setIsLoading(false);
-    };
-    checkAuth();
-  }, []);
+    if (import.meta.env.VITE_STORAGE_MODE !== "remote") {
+      let worker: Worker | null = null;
 
-  useEffect(() => {
-    let worker: Worker | null = null;
-
-    if (identifier) {
-      worker = new Worker(new URL("./workers/scheduler.ts", import.meta.url));
-      worker.postMessage("start");
-    }
-
-    return () => {
-      if (worker) {
-        worker.postMessage("stop");
-        worker.terminate();
+      if (identifier) {
+        worker = new Worker(new URL("./workers/scheduler.ts", import.meta.url));
+        const credentials = localStorage.getItem("apiCredentials");
+        worker.postMessage({
+          type: "start",
+          credentials,
+        });
       }
-    };
+
+      return () => {
+        if (worker) {
+          worker.postMessage("stop");
+          worker.terminate();
+        }
+      };
+    }
   }, [identifier]);
 
   if (isLoading) {
@@ -44,18 +49,36 @@ function App() {
     );
   }
 
+  if (import.meta.env.VITE_STORAGE_MODE === "remote" && !hasApiCredentials) {
+    return (
+      <div className="relative">
+        <SetupForm onSuccess={() => updateApiAuth(true)} />
+        <Toaster position="top-right" />
+      </div>
+    );
+  }
+
+  if (import.meta.env.VITE_STORAGE_MODE === "remote" && !isApiAuthenticated) {
+    return (
+      <div className="relative">
+        <ApiLoginForm onSuccess={() => updateApiAuth(true)} />
+        <Toaster position="top-right" />
+      </div>
+    );
+  }
+
   if (!identifier) {
     return (
       <div className="relative">
-        <LoginForm onSuccess={setIdentifier} />
+        <LoginForm onSuccess={updateIdentifier} />
         <Toaster position="top-right" />
       </div>
     );
   }
 
   const handleLogout = async () => {
-    await db.credentials.clear(); // Clear credentials from the database
-    setIdentifier(undefined); // Update authenticated state
+    await db.deleteCredentials();
+    updateIdentifier(undefined);
   };
 
   return (
@@ -72,7 +95,7 @@ function App() {
       </header>
 
       <main className="px-6 lg:px-8 mb-8">
-        <div className="max-w-7xl  mx-auto px-4 py-6 sm:px-0 sm:grid sm:grid-cols-2 gap-2">
+        <div className="max-w-7xl mx-auto px-4 py-6 sm:px-0 sm:grid sm:grid-cols-2 gap-2">
           <div className="w-full">
             <PostScheduler />
           </div>
